requirements.txt
mlflow
pandas
scikit-learn
numpy
matplotlib
seaborn
scipy
apache-airflow
psycopg2-binary
acryl-datahub[airflow]
fastapi
streamlit
joblib
uvicorn
pydantic
typing
dill
python-multipart